src/lambda/requirements.txt
pandas==1.5.3
psycopg2-binary==2.9.9
boto3==1.28.44